// @types/FiiTijoloCalculatedDataType.d.ts
export type FiiTijoloCalculatedDataType = {
  ticker: string;
  category: string;
  price: number;
  dy: number;
  pvp: number;
  riskPremium: number;
  discountRate: number;
  growthRate: number;
  dividendYear1: number;
  presentValue1: number;
  dividendYear2: number;
  presentValue2: number;
  dividendYear3: number;
  presentValue3: number;
  dividendYear4: number;
  presentValue4: number;
  dividendYear5: number;
  presentValue5: number;
  dividendYear6: number;
  presentValue6: number;
  dividendYear7: number;
  presentValue7: number;
  dividendYear8: number;
  presentValue8: number;
  dividendYear9: number;
  presentValue9: number;
  dividendYear10: number;
  presentValue10: number;
  desinvestment: number;
  presentValueDesinvestment: number;
  fairPrice: number;
  ceelingPrice: number;
  expectativaCrescimento: number;
  caixa: number;
  cagrDividendos3Anos: number;
  cagrValorCota3Anos: number;
  gestao: string;
  isTopManager: boolean;
  patrimonio: number;
  qtdImoveis: number;
  ativos: string;
  locatario: string;
  cotistas: number;
  liquidezDiaria: number;
  precoM2: number;
  aluguelM2: number;
  capRate: number;
  vacanciaMedia: number;
};